requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
motor
python-jose
dnspython
python-dotenv
PyJWT